requirements.txt
blinker==1.7.0
click==8.1.7
flask==3.0.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
numpy==1.26.3
opencv-python==4.9.0.80
PyMySQL==1.1.0
RPi.GPIO==0.7.1
werkzeug==3.0.1
zipp==3.17.0
